Added preliminary OpenCl BIP39 support (Not very fast, unsupported)
After looking at a few different implementations (Hashcat, John the Ripper) I have just gone with opencl_brute, as it's already in Python and easy to work with. (As well as to add OpenCL support to other walelt types)

In terms of BIP39 and Electrum, it is currently heavily CPU bound, but at least works for now. (Still gives a 50% performance improvement over my CPU) Will look at some ways to optimise this. (The pyopencl_brute library seems to give about 10x CPU performance, not the 100x of hashcat, but until it is sufficiently non-cpu bound to max out the current kernel, it isn't worth optimising on the opencl side)

BIP39 performance will never really be able to reach hashcat (or wallet file) levels of performance, as they are simply brute-forcing a short string (wallet extract) and looking for a similarly short string as the output. BIP39 brute-forcing is different in that going from seed -> address requires several additioal steps of single hashing, etc, that significantly slows things down. (This can be easily seen if running as one thread)