Add sensible auto-chunksize settings and add ability to do checksum in generator
A few changes to make OpenCL processing more useful.
*Add sensible auto worksize calculations if a manual workgroup size isn't provided. (These aim to ensure that the max workgroup of the GPU is fully utilised for every chunk)
*Standardise arguments a bit. Remove ones that aren't used, match arguments to information displayed in OpenCl dialogues and standardise around OpenCl language as opposed to just GPU.
*Add ability to do checksumming in the generator as opposed to worker threads. I was curious to see whether this would give much of a performance boost.  (It doesn't really, quickly bottleknecks as the generator runs in a single thread)

As it stands, the performance benefit is tied to where the it was bottleknecking when compared to pure CPU. A BIP39 12 word seed checksum is very weak and only weeds out 1/16 of the candidate seeds, so it sees about a 4x performance increase vs CPU. An Electrum or 24 word seed sees a much more modest boost due to most seeds being weeded out before hashing due to the stronger checksumming. Both are around 1.5 -> 2x faster than pure CPU. Still quite worthwhile.