Add more flexibility to restoring autosave files
Have added a list of parameters which can be savely added/removed while still preserving the ability to use an autosave file.

This is handy if you want to change things like --no-dupchecks if you were running out of memory, or start a process on CPU and then move to GPU... change thread count, etc...

Processing of parameters are indicated in the program output when an autosave file is used.